feat: load participant code as JS modules instead of iframe html
The old way will still work, so it is a backwards compatible change.
However, it is now recommended to use the participant index.js to export a template (supports lit-html too, as well as exporting a HTML Node) which will be rendered by the CWK App Shell.
Furthermore, it is recommended to then also have an index.html which imports this template or node, and renders it, so that the participant can click "View" and see only their code.

This is now all documented in the main README