feat: supports only a single 'dir' arg, support iframe configurations now that we allow modules
BREAKING CHANGE: For a short time, specifying app-index through appIndex (--app-index) was allowed, but to keep things simple, you are only allowed to specify the directory to run CWK on, which should contain the index.html if you want a CWK App Shell, as well as the cwk.config.js. The option is called 'dir' or --dir in the CLI.
BREAKING CHANGE: To keep things simpler, the --output-dir, --input-dir and --config-dir options have been removed from the scaffold command. To keep things simple, just specify --dir to your directory which should contain a template folder, and the output is always a participants folder. This keeps things a lot simpler than they were before, and prevents having to solve a bunch of problems that showed up when this was flexible.