feat: POC allow passing cmd before rerun, which runs sync, for cleanup
Useful if you have a child that spawns an http-server that runs as its own process which doesn't get killed properly when rerunning.
This allows you to clean it up manually. It's just a workaround. I intend to find a proper fix so that on re-run, child processes get killed properly..